ml-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.0
nltk==3.8.1
textblob==0.17.1